+++ src/mqtt/mod.rs
@@ -1,1 +1,2 @@
 pub mod mqtt;
+pub mod mqttsprinkler;
